Update README.md.AllowFundingInETHfrom MultipleAddresses
This is a smart contract developed in Solidity for the Ethereum blockchain. It is designed to allow funding in Ether (ETH) from multiple addresses, with each funding operation subject to a minimum amount in USD. The conversion from ETH to USD is done using Chainlink's AggregatorV3Interface, which is a decentralized oracle network providing reliable and tamper-proof inputs and outputs for complex smart contracts on any blockchain.

Key Components
AggregatorV3Interface: This is a Chainlink interface used to get the latest price data. The contract is initialized with an address pointing to a Chainlink price feed.

MinimumUSD: This is a public variable that defines the minimum amount of USD equivalent that can be funded.

Funders and addressToAmountFunded: These are a dynamic array and a mapping used to keep track of funders and the amount each has funded, respectively.

Core Functions
fund: This function allows a user to fund the contract. It checks whether the transferred ETH amount is greater than or equal to the minimumUSD. If the check passes, it records the sender's address and the funded amount.

withdraw: This function is used to withdraw funds. The function currently loops over all funders, but the withdrawal logic needs to be added.

getPrice: This function returns the latest price data (ETH to USD conversion rate) using the Chainlink's AggregatorV3Interface.

getConversionRate: This function takes an amount in ETH and returns its equivalent value in USD, using the latest price data.